fix: fix `make generate` to build with `BUILDKIT_MULTI_PLATFORM=0`
Due to the change in `BUILDKIT_MULTI_PLATFORM=1`, the generated files from `make generate` land initially in a directory named `<os>_<arch>`, which we then move into the actual destination directories with a script.

This move fails to merge the existing directories with the newly generated files. So, instead, on `generate` command, we override the `BUILDKIT_MULTI_PLATFORM` variable back to `0` to restore the old behavior. We do not need it for generating sources.

Additionally, remove the deprecated `gomnd` linter from the `golangci-lint` configuration.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>