fix: make JS frontend target attach to the build
Previously it was attached to the `base`, so it got into steps like
`golangci-lint` which don't care about it, but also it gets exported
back to the source tree via `make lint-golangci-lint-fmt`.

Re-attach it to the build step (where we actually need it).

Another solution is to make it part of `make generate` and keep it part
of the source tree.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>